[ELK] Rename finos_meetings to finosmeetings
For consistency reason with the corresponding perceval backend,
the connectors have been renamed finosmeetings.
